FilterReducer
