+++ src/actions/__tests__/range-test.js
@@ -3,7 +3,7 @@ import getRange from '../range'
 describe('actions/getRange', () => {
   it('returns range between 2 and totalPage if totalPage' +
      'is larger than preset range', () => {
-    expect(getRange(1, 2, 3)).toEqual([2])
+    expect(getRange(1, 1, 3)).toEqual([2])
   })
 
   it('returns range before and after current number according to set limit', () => {
