+++ src/reducers/__tests__/queryReducer-test.js
@@ -6,18 +6,10 @@ const { Map } = require('immutable')
 describe('queryReducer', () => {
   const data = [1, 2, 3]
 
-  const filter = index => Map({
-    type: `type-${index}`,
-    availableValues: [`value-${index}`],
-    value: index
-  })
-
   const initialState = Map({
-    categories: [],
     keyword: '',
     offset: 0,
     pageNum: 1,
-    filters: [filter(0), filter(1)],
     data: undefined,
     error: []
   })
@@ -26,11 +18,9 @@ describe('queryReducer', () => {
     expect(
      queryReducer(undefined, {})
    ).toEqual(Map({
-     categories: [],
      keyword: '',
      offset: 0,
      pageNum: 1,
-     filters: [],
      data: undefined,
      error: []
    }))
@@ -43,11 +33,9 @@ describe('queryReducer', () => {
         data
       })
     ).toEqual(Map({
-      categories: [],
       keyword: '',
       offset: 0,
       pageNum: 1,
-      filters: [filter(0), filter(1)],
       data,
       error: []
     }))
@@ -60,11 +48,9 @@ describe('queryReducer', () => {
         error: 'Some Error'
       })
     ).toEqual(Map({
-      categories: [],
       keyword: '',
       offset: 0,
       pageNum: 1,
-      filters: [filter(0), filter(1)],
       data: undefined,
       error: 'Some Error'
     }))
@@ -77,114 +63,9 @@ describe('queryReducer', () => {
         keyword: 'Some Keyword'
       })
     ).toEqual(Map({
-      categories: [],
       keyword: 'Some Keyword',
       offset: 0,
       pageNum: 1,
-      filters: [filter(0), filter(1)],
-      data: undefined,
-      error: []
-    }))
-  })
-
-  it('handles add filter', () => {
-    expect(
-      queryReducer(initialState, {
-        type: actionTypes.ADD_FILTER
-      })
-    ).toEqual(Map({
-      categories: [],
-      keyword: '',
-      offset: 0,
-      pageNum: 1,
-      filters: [filter(0), filter(1), Map({
-        type: '',
-        availableValues: [],
-        value: ''
-      })],
-      data: undefined,
-      error: []
-    }))
-  })
-
-  it('handles remove selected filter', () => {
-    expect(
-      queryReducer(initialState, {
-        type: actionTypes.REMOVE_SELECTED_FILTER,
-        index: 1
-      })
-    ).toEqual(Map({
-      categories: [],
-      keyword: '',
-      offset: 0,
-      pageNum: 1,
-      filters: [filter(0)],
-      data: undefined,
-      error: []
-    }))
-  })
-
-  it('handles remove all filters', () => {
-    expect(
-      queryReducer(initialState, {
-        type: actionTypes.REMOVE_ALL_FILTERS
-      })
-    ).toEqual(Map({
-      categories: [],
-      keyword: '',
-      offset: 0,
-      pageNum: 1,
-      filters: [],
-      data: undefined,
-      error: []
-    }))
-  })
-
-  it('handles update selected filter', () => {
-    expect(
-      queryReducer(initialState, {
-        type: actionTypes.UPDATE_SELECTED_FILTER,
-        index: 1,
-        selectedFilter: 'type-1'
-      })
-    ).toEqual(Map({
-      categories: [],
-      keyword: '',
-      offset: 0,
-      pageNum: 1,
-      filters: [
-        filter(0),
-        Map({
-          type: 'type-1',
-          availableValues: [],
-          value: ''
-        })
-      ],
-      data: undefined,
-      error: []
-    }))
-  })
-
-  it('handles update selected filter value ', () => {
-    expect(
-      queryReducer(initialState, {
-        type: actionTypes.UPDATE_SELECTED_FILTER_VALUE,
-        index: 1,
-        selectedValue: '1'
-      })
-    ).toEqual(Map({
-      categories: [],
-      keyword: '',
-      offset: 0,
-      pageNum: 1,
-      filters: [
-        filter(0),
-        Map({
-          type: 'type-1',
-          availableValues: ['value-1'],
-          value: '1'
-        })
-      ],
       data: undefined,
       error: []
     }))
@@ -197,28 +78,25 @@ describe('queryReducer', () => {
         pageNum: 2
       })
     ).toEqual(Map({
-      categories: [],
       keyword: '',
       offset: 10,
       pageNum: 2,
-      filters: [filter(0), filter(1)],
       data: undefined,
       error: []
     }))
   })
 
-  it('handles update categories', () => {
+  it('clears error', () => {
+    initialState.set('error', ['Some Error'])
+
     expect(
       queryReducer(initialState, {
-        type: actionTypes.UPDATE_CATEGORIES,
-        categories: ['1', '2', '3']
+        type: actionTypes.CLEAR_ERROR
       })
     ).toEqual(Map({
-      categories: ['1', '2', '3'],
       keyword: '',
       offset: 0,
       pageNum: 1,
-      filters: [filter(0), filter(1)],
       data: undefined,
       error: []
     }))
