+++ src/reducers/index.js
@@ -2,11 +2,13 @@ import { combineReducers } from 'redux'
 import { reducer as formReducer } from 'redux-form'
 import queryReducer from './queryReducer'
 import fetchingReducer from './fetchingReducer'
+import filterReducer from './filterReducer'
 
 const rootReducer = combineReducers({
   form: formReducer,
   query: queryReducer,
-  isFetching: fetchingReducer
+  isFetching: fetchingReducer,
+  filters: filterReducer
 })
 
 export default rootReducer
